Add ajax "page"
An overarching renderer that runs the page itself in an ErrorT monad, so that it
can report errors in an intelligent way, which can then be inserted into the
Javelin slot for ajax error responses.

This took me way too long to figure out. It's obvious in hindsight, but that
didn't save me from an evening of flailing through monad transformer hell,
getting C++-esque error messages, thinking that you really do need to be a
category theorist to write Haskell. But on the flipside, once my code
typechecked, it did work the first try. :)

For the curious, example error message just from a forgotten "lift":

Page/Delete.hs:23:18:
    No instance for (S.HasRqData (E.ErrorT String (S.ServerPartT IO)))
      arising from a use of `S.lookText'
    Possible fix:
      add an instance declaration for
      (S.HasRqData (E.ErrorT String (S.ServerPartT IO)))
    In a stmt of a 'do' block: idstr <- S.lookText "id"
    In the expression:
      do { idstr <- S.lookText "id";
           id <- case R.decimal idstr of {
                   Left err -> lift $ E.throwError err
                   Right (i, _) -> return i };
           updateResult <- AA.update' acid $ IL.DeleteItem $ I.ItemID id;
           () <- case updateResult of {
                   Left err -> lift $ E.throwError err
                   Right () -> return () };
           .... }
    In an equation for `render':
        render acid
          = do { idstr <- S.lookText "id";
                 id <- case R.decimal idstr of {
                         Left err -> lift $ E.throwError err
                         Right (i, _) -> return i };
                 updateResult <- AA.update' acid $ IL.DeleteItem $ I.ItemID id;
                 .... }